requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.5.0
pydantic==2.6.4
sqlalchemy==2.0.29
psycopg[binary]==3.1.19
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
bcrypt==3.2.0
